Доработка интерфейса

+++ Platform/IRenderer.cs
@@ -6,7 +6,7 @@ public interface IRenderer
 
     void BeginFrame();
     void EndFrame();
-    void Clear(float r, float g, float b, float a);
+    void Clear(int r, int g, int b, int a);
 
     ITexture CreateTexture(string path);
     ITexture CreateTexture(int width, int height, byte[] data);
